+++ src/register-component/index.js
@@ -1,0 +1,3 @@
+import registerComponent from './register-component';
+
+module.exports = registerComponent;